+++ src/styles/_theme.scss
@@ -26,6 +26,7 @@
   --architect-text: var(--color-charcoal);
   --architect-shadow: var(--color-shadow);
   --architect-divider: var(--color-platinum);
+  --architect-selected: var(--color-sea-green);
 
   // Control bar
   --architect-control-bar-background: var(--color-cyber-grape);
